LayoutTests/http/tests/security/frameNavigation/navigation-DENIED-sibling-form-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to initiate a navigation change for frame with URL http://127.0.0.1:8000/security/frameNavigation/resources/ready.html from frame with URL http://localhost:8000/security/frameNavigation/resources/ready.html.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to initiate a navigation change for frame with URL http://127.0.0.1:8000/security/frameNavigation/resources/ready.html from frame with URL http://localhost:8000/security/frameNavigation/resources/ready.html.

Navigating frame.
Window count: 2
Done.
 

--------
Frame: 'child2'
--------
127.0.0.1:8000

Waiting for navigation commands.

--------
Frame: 'active2'
--------
localhost:8000

Waiting for navigation commands.
